backend/requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asyncpg==0.30.0
attrs==25.1.0
Authlib==1.6.0
bcrypt==4.3.0
bidict==0.23.1
bitarray==3.2.0
bitstring==4.3.1
blinker==1.9.0
branca==0.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cryptography==45.0.4
cycler==0.12.1
dataclasses-json==0.6.7
deepgram-sdk==2.12.0
deprecation==2.1.0
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
esptool==4.8.1
eventlet==0.40.4
fastapi==0.111.0
fastapi-cli==0.0.11
filelock==3.18.0
Flask==3.1.0
flask-cors==6.0.2
Flask-Mail==0.10.0
Flask-SocketIO==5.5.1
folium==0.20.0
fonttools==4.59.2
fpdf==1.7.2
frozenlist==1.5.0
fsspec==2025.5.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
imutils==0.5.4
intelhex==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.3
kiwisolver==1.4.9
llvmlite==0.44.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
msal==1.32.3
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==2.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.4
oauthlib==3.2.2
openai==1.97.0
openai-whisper==20240930
opencv-python==4.11.0.86
orjson==3.10.7
packaging==24.2
pandas==2.3.1
passlib==1.7.4
pillow==11.2.1
plotly==6.3.0
propcache==0.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.9.1
pydantic_core==2.20.1
Pygments==2.19.2
PyJWT==2.9.0
pymongo==4.15.0
pyparsing==3.2.3
pyserial==3.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.3
python-jose==3.5.0
python-multipart==0.0.9
python-socketio==5.15.0
pytz==2025.2
PyYAML==6.0.2
razorpay==1.4.2
reedsolo==1.7.0
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
rich-toolkit==0.15.1
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.16.3
secure-smtplib==0.1.1
setuptools==80.9.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.37.2
stripe==12.2.0
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.9.0
torch==2.7.0+cpu
torchaudio==2.7.0+cpu
torchvision==0.22.0+cpu
tqdm==4.67.1
typer==0.17.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.4.0
uvicorn==0.30.1
waitress==3.0.2
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.3.2
xyzservices==2025.4.0
yarl==1.18.3
